requirements.txt
requests>=2.28.0
PyYAML>=6.0
python-pptx>=0.6.21
python-docx>=0.8.11
Jinja2>=3.1.0
Faker>=24.0.0
yake>=0.4.8
markdown>=3.5.0
